requirements.txt
Django==4.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.6
coreapi==2.3.3
djangorestframework-simplejwt==5.2.2
coverage==7.2.7
requests==2.31.0
celery==5.3.1
redis==4.6.0
django-celery-beat==2.5.0
aiogram==2.25.1
django-cors-headers==4.2.0
gunicorn
whitenoise
flake8==6.0.0
